refactor: merge KV into R2D2
