--- packages/web3-providers/src/KV/README.md
@@ -1,16 +1,0 @@
-# Key-Value storage
-
-```typescript
-// create a JSON storage
-const storage = KeyValue.createJSON_Storage('hello_world')
-
-// set value
-await storage.set('foo', {
-  key: 'value',
-})
-
-// get value
-const response = await storage.get('foo')
-
-console.assert(response?.key === 'value', 'ASSERTION FAILED')
-```
