--- packages/web3-providers/src/KV/constants.ts
@@ -1,1 +1,0 @@
-export const KV_ROOT_URL = 'https://kv.r2d2.to'
