+++ packages/web3-providers/src/R2D2/constants.ts
@@ -1,1 +1,2 @@
-export const API_URL = 'https://ens-reverse-lookup.r2d2.to/ens/'
+export const KV_ROOT_URL = 'https://kv.r2d2.to'
+export const ENS_ROOT_URL = 'https://ens-reverse-lookup.r2d2.to/ens/'
