+++ packages/web3-providers/src/R2D2/index.ts
@@ -1,2 +1,3 @@
 export * from './apis/TokenListAPI.js'
 export * from './apis/DomainAPI.js'
+export * from './apis/StorageAPI.js'
